+++ src/main/kotlin/com/github/jpvos/keylogger/plugin/listeners/KeyloggerListener.kt
@@ -2,7 +2,7 @@ package com.github.jpvos.keylogger.plugin.listeners
 
 import com.github.jpvos.keylogger.core.Action
 import com.github.jpvos.keylogger.plugin.KeyloggerBundle
-import com.github.jpvos.keylogger.plugin.services.KeyloggerService
+import com.github.jpvos.keylogger.plugin.services.CounterService
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.AnActionResult
@@ -13,10 +13,10 @@ import com.intellij.openapi.editor.event.EditorMouseEvent
 import com.intellij.openapi.editor.event.EditorMouseListener
 
 internal class KeyloggerListener : AnActionListener, EditorMouseListener {
-    private val keyloggerService = service<KeyloggerService>()
+    private val counterService = service<CounterService>()
     override fun afterActionPerformed(action: AnAction, event: AnActionEvent, result: AnActionResult) {
         super.afterActionPerformed(action, event, result)
-        keyloggerService.counter.next(
+        counterService.counter.next(
             Action(
                 Action.Type.EditorAction,
                 action.templatePresentation.text?.toString() ?: KeyloggerBundle.message("general.unknown")
@@ -26,7 +26,7 @@ internal class KeyloggerListener : AnActionListener, EditorMouseListener {
 
     override fun afterEditorTyping(c: Char, dataContext: DataContext) {
         super.afterEditorTyping(c, dataContext)
-        keyloggerService.counter.next(
+        counterService.counter.next(
             Action(
                 Action.Type.Keystroke,
                 "<$c>"
@@ -36,7 +36,7 @@ internal class KeyloggerListener : AnActionListener, EditorMouseListener {
 
     override fun mousePressed(event: EditorMouseEvent) {
         super.mousePressed(event)
-        keyloggerService.counter.next(
+        counterService.counter.next(
             Action(
                 Action.Type.Mouse,
                 KeyloggerBundle.message("general.click")