+++ src/main/kotlin/com/github/jpvos/keylogger/plugin/services/DatabaseService.kt
@@ -1,67 +1,40 @@
 package com.github.jpvos.keylogger.plugin.services
 
 import com.github.jpvos.keylogger.core.Action
-import com.github.jpvos.keylogger.core.Counter
-import com.github.jpvos.keylogger.db.DatabaseConnection
+import com.github.jpvos.keylogger.db.SqliteDatabaseConnection
+import com.github.jpvos.keylogger.plugin.settings.KeyloggerSettings
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.components.Service
-import com.intellij.openapi.diagnostic.thisLogger
-import kotlin.io.path.Path
 
-
-// TODO: does Disposable work correctly for this service?
 @Service
-class KeyloggerService : Counter.Listener, Disposable {
-    private val filePath = Path(System.getProperty("user.home"), "intellij-plugin-keylogger.sqlite") // TODO: configurable
-    private val idleTimeout = 1000L // TODO: configurable
-    private val db = DatabaseConnection("jdbc:sqlite:file:$filePath")
-    private val logger = thisLogger()
-    val counter = Counter(idleTimeout = idleTimeout)
+class DatabaseService : Disposable {
+    private var db = SqliteDatabaseConnection()
 
     init {
-        initializeDatabase()
-        loadInitialState()
-        counter.registerListener(this)
+        restoreDatabase()
     }
 
-
     override fun dispose() {
-        counter.unregisterListener(this)
         db.close()
     }
 
-    override fun onAction(action: Action) {
-        db.preparedStatement("INSERT INTO actions (type, name, timestamp) VALUES (?, ?, ?)") {
-            setString(1, action.type.name)
-            setString(2, action.name)
-            setLong(3, System.currentTimeMillis())
-            execute()
-        }
-    }
-
-    private fun initializeDatabase() {
-        logger.warn("Initializing Keylogger database")
+    fun restoreDatabase() {
+        dispose()
+        db.connect(KeyloggerSettings.instance.databaseURL)
         db.statement("CREATE TABLE IF NOT EXISTS meta (key TEXT NOT NULL UNIQUE, value TEXT)")
         val pluginVersion = db.query("SELECT value FROM meta WHERE key = 'plugin_version'") {
             if (next()) getString("value") else null
         }
-        pluginVersion?.let {
-            logger.warn("Plugin database version $it detected")
-        } ?: run {
-            logger.warn("No plugin database version found, create tables")
+        // in the future, use this plugin version number to implement database migrations
+        // for now, just create the tables if they don't exist
+        if (pluginVersion == null) {
             db.statement("INSERT INTO meta (key, value) VALUES ('plugin_version', '1')")
             db.statement("CREATE TABLE IF NOT EXISTS actions (type TEXT NOT NULL, name TEXT NOT NULL, timestamp LONG NOT NULL)")
         }
-        logger.warn("Keylogger database initialized")
     }
 
-    private fun loadInitialState() {
-        val actions = readActionsFromDatabase()
-        val (activeTime, idleTime) = readActiveAndIdleTimeFromDatabase()
-        counter.setState(Counter.State(actions, activeTime, idleTime))
-    }
-
-    private fun readActionsFromDatabase(): Map<Action, Long> {
+    fun queryActionsMap(): Map<Action, Long> {
+        verifyDatabaseConnection()
         return db.query("SELECT type, name, count(*) as count FROM actions GROUP BY type, name") {
             val result = mutableMapOf<Action, Long>()
             while (next()) {
@@ -81,7 +54,9 @@ class KeyloggerService : Counter.Listener, Disposable {
         }
     }
 
-    private fun readActiveAndIdleTimeFromDatabase(): Pair<Long, Long> {
+    fun queryActiveAndIdleTime(): Pair<Long, Long> {
+        verifyDatabaseConnection()
+        // this implementation is not efficient, but it is good enough for now. will it stand the test of time?
         return db.query("SELECT timestamp FROM actions ORDER BY timestamp ASC") {
             var currentTime: Long? = null
             var activeTime = 0L
@@ -92,7 +67,7 @@ class KeyloggerService : Counter.Listener, Disposable {
                     currentTime = timestamp
                 }
                 val deltaTime = timestamp - currentTime
-                if (deltaTime < idleTimeout) {
+                if (deltaTime < KeyloggerSettings.instance.idleTimeout) {
                     activeTime += deltaTime
                 } else {
                     idleTime += deltaTime
@@ -101,6 +76,20 @@ class KeyloggerService : Counter.Listener, Disposable {
             }
             return@query activeTime to idleTime
         }
+    }
+
+    fun persistAction(action: Action) {
+        verifyDatabaseConnection()
+        db.preparedStatement("INSERT INTO actions (type, name, timestamp) VALUES (?, ?, ?)") {
+            setString(1, action.type.name)
+            setString(2, action.name)
+            setLong(3, System.currentTimeMillis())
+            execute()
+        }
+    }
+
+    private fun verifyDatabaseConnection() {
+        if (!db.open) throw Error("Database is not open, initialize database first")
     }
 
 }