+++ src/main/kotlin/com/github/jpvos/keylogger/plugin/toolwindow/components/StatisticsComponent.kt
@@ -1,18 +1,19 @@
-package com.github.jpvos.keylogger.plugin.toolwindow.fragments
+package com.github.jpvos.keylogger.plugin.toolwindow.components
 
 import com.github.jpvos.keylogger.core.Action
 import com.github.jpvos.keylogger.core.Counter
 import com.github.jpvos.keylogger.plugin.KeyloggerBundle
-import com.github.jpvos.keylogger.plugin.services.KeyloggerService
+import com.github.jpvos.keylogger.plugin.services.CounterService
+import com.github.jpvos.keylogger.plugin.settings.KeyloggerSettings
 import com.github.jpvos.keylogger.ui.Container
 import com.github.jpvos.keylogger.ui.DisplayFormat
 import com.github.jpvos.keylogger.ui.Table
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.components.service
 
-class StatisticsTable : Container(), Counter.Listener, Disposable {
+class StatisticsComponent : Container(), Counter.Listener, Disposable {
 
-    private val keyloggerService = service<KeyloggerService>()
+    private val counterService = service<CounterService>()
     private val table = Table(
         arrayOf(
             KeyloggerBundle.message("statistics.table.information"),
@@ -23,7 +24,7 @@ class StatisticsTable : Container(), Counter.Listener, Disposable {
     init {
         add(table)
         updateTableData()
-        keyloggerService.counter.registerListener(this)
+        counterService.counter.registerListener(this)
     }
 
     override fun onAction(action: Action) {
@@ -31,11 +32,11 @@ class StatisticsTable : Container(), Counter.Listener, Disposable {
     }
 
     override fun dispose() {
-        keyloggerService.counter.unregisterListener(this)
+        counterService.counter.unregisterListener(this)
     }
 
     private fun updateTableData() {
-        val state = keyloggerService.counter.getState()
+        val state = counterService.counter.getState()
         table.setTableData(
             arrayOf(
                 arrayOf(
@@ -48,15 +49,21 @@ class StatisticsTable : Container(), Counter.Listener, Disposable {
                 ),
                 arrayOf(
                     KeyloggerBundle.message("statistics.totalTime"),
-                    DisplayFormat.milliseconds(state.totalTime)
+                    DisplayFormat.duration(state.totalTime)
                 ),
                 arrayOf(
                     KeyloggerBundle.message("statistics.activeTime"),
-                    DisplayFormat.milliseconds(state.activeTime)
+                    DisplayFormat.duration(state.activeTime)
                 ),
                 arrayOf(
                     KeyloggerBundle.message("statistics.idleTime"),
-                    DisplayFormat.milliseconds(state.idleTime)
+                    DisplayFormat.duration(state.idleTime)
+                ),
+
+                arrayOf(
+                    KeyloggerBundle.message("statistics.idleTimeout"),
+                    DisplayFormat.milliseconds(KeyloggerSettings.instance.idleTimeout)
+
                 ),
                 arrayOf(
                     KeyloggerBundle.message("statistics.actionsPerMinute"),