+++ src/main/kotlin/com/github/jpvos/keylogger/ui/Tabs.kt
@@ -1,7 +1,6 @@
 package com.github.jpvos.keylogger.ui
 
 import com.intellij.ui.components.JBTabbedPane
-import java.awt.event.KeyEvent
 import javax.swing.JComponent
 
 class Tabs : Container() {
@@ -15,7 +14,6 @@ class Tabs : Container() {
 
     fun addTab(title: String, content: JComponent) {
         tabbedPane.addTab(title, null, content, null)
-        tabbedPane.setMnemonicAt(size, KeyEvent.VK_1 + size) // TODO: doesnt seem to work
         size++
     }
 }